requirements.txt
aiofiles==23.2.1
aiogram==3.13.0
aiogram_dialog==2.2.0
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
amqp==5.2.0
annotated-types==0.7.0
asgiref==3.8.1
attrs==24.2.0
billiard==4.2.0
cachetools==5.5.0
celery==5.4.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.5
Django==5.1.1
django-celery-beat==2.7.0
django-timezone-field==7.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
frozenlist==1.4.1
idna==3.10
isort==5.13.2
Jinja2==3.1.4
kombu==5.4.1
magic-filter==1.0.12
MarkupSafe==2.1.5
multidict==6.1.0
prompt_toolkit==3.0.47
pydantic==2.8.2
pydantic_core==2.20.1
PyJWT==2.9.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
redis==5.0.8
requests==2.32.3
six==1.16.0
sqlparse==0.5.1
tree-cli==0.1.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.13
yarl==1.11.1
psycopg2-binary==2.9.9